sbn: remove `sbn_free()` and `sbn_clone()`
Don't want to encourage unnecessary memory usage, since we can allocate
sbns on the stack. The struct is also so simple that allocating and
freeing one is not too cumbersome, in cases one really needs it (just
`[mc]alloc()` and `free()`); copying the contents of one struct to
another is just a matter `*b = *a`.